wrote more aStar
completed A* steps 1-5.  Still needs a beter heuristic for H or the estimated distance from the tile to the target, and needs an algorithm to loop back through the closed list to get the shortest path(choose lowest totalCost value maybe?  But I think that is just dkikstra, unless the open list H hueristic is really that good.)